Add Hono on Bun to the HTTP routing benchmark
Bun is the fastest JS runtime for Hono, so it serves as a "fastest JS"
reference point alongside Hono on Node.js. Hono route definitions are
factored into a shared app.routes.js so the Node and Bun entry points
run identical routing logic.
