+++ benchmark/README.md
@@ -143,29 +143,30 @@ Notes:
 ## HTTP Routing
 
 End-to-end HTTP throughput of `wado serve` vs an equivalent
-[Hono](https://hono.dev/) server on Node.js vs an equivalent
-[Axum](https://github.com/tokio-rs/axum) server in native Rust. The
-route and request set is Hono's own official router benchmark
+[Hono](https://hono.dev/) server on Node.js and on Bun, vs an
+equivalent [Axum](https://github.com/tokio-rs/axum) server in native
+Rust. The route and request set is Hono's own official router benchmark
 (`honojs/hono`, `benchmarks/routers/`), driven over HTTP with `oha`
 (6s, 50 connections per request).
 
 Throughput (requests/sec, higher is better):
 
-| Request                         | `wado serve` | Hono (Node) | Axum (native) |
-| ------------------------------- | -----------: | ----------: | ------------: |
-| `GET /user`                     |       30,728 |      25,701 |       138,236 |
-| `GET /user/lookup/username/hey` |       29,608 |      21,466 |       119,565 |
-| `GET /event/abcd1234/comments`  |       28,862 |      24,815 |       124,284 |
-| `POST /event/abcd1234/comment`  |       28,721 |      18,772 |       132,484 |
-| `GET /static/index.html`        |       30,326 |      20,907 |       119,672 |
+| Request                         | `wado serve` | Hono (Node) | Hono (Bun) | Axum (native) |
+| ------------------------------- | -----------: | ----------: | ---------: | ------------: |
+| `GET /user`                     |       37,344 |      32,808 |     51,510 |       134,636 |
+| `GET /user/lookup/username/hey` |       35,882 |      34,372 |     48,369 |       118,892 |
+| `GET /event/abcd1234/comments`  |       36,187 |      34,157 |     47,926 |       117,129 |
+| `POST /event/abcd1234/comment`  |       37,753 |      31,106 |     45,849 |       114,574 |
+| `GET /static/index.html`        |       40,750 |      31,091 |     48,984 |       114,157 |
 
-`wado serve` leads Hono on every request (~29k–30k vs ~19k–28k req/s);
-native-Rust Axum is the ceiling at ~4–5x faster (~120k–138k). `wado
-serve` runs a `wasi:http/service` component on wasmtime, dispatching
-through `core:router`, with pooled instance reuse + periodic recycling.
-A cross-runtime comparison (Wasm component on wasmtime vs JS on Node.js
-vs native Rust). See `http_routing/README.md` for the full table and
-methodology.
+`wado serve` leads Hono on Node on every request (~36k–41k vs ~31k–36k
+req/s) but trails Hono on Bun (~46k–52k) — Bun's HTTP server is
+markedly faster than Node's. Native-Rust Axum is the ceiling at ~3x
+faster (~114k–135k). `wado serve` runs a `wasi:http/service` component
+on wasmtime, dispatching through `core:router`, with pooled instance
+reuse + periodic recycling. A cross-runtime comparison (Wasm component
+on wasmtime vs JS on Node.js/Bun vs native Rust). See
+`http_routing/README.md` for the full table and methodology.
 
 ## Running
