+++ benchmark/mise.toml
@@ -19,6 +19,7 @@ lockfile = true
 
 [tools]
 node = "latest"
+bun = "latest"
 zig = "latest"
 # "github:WebAssembly/wasi-sdk" = "latest" # clang + wasm-ld + wasi-sysroot for C zlib Wasm (disabled: GH rate limit)
 
@@ -268,7 +269,7 @@ cargo run --release --manifest-path ../wado-cli/Cargo.toml --quiet -- run -O2 sy
 """
 
 [tasks.http-routing]
-description = "Run HTTP routing benchmark (wado serve vs Hono on Node.js)"
+description = "Run HTTP routing benchmark (wado serve vs Hono on Node.js and Bun vs Axum)"
 run = "./http_routing/bench.sh"
 
 [tasks.all]
