+++ mise.toml
@@ -453,7 +453,7 @@ description = "Run syntax-highlight benchmark"
 run = "mise run -C benchmark syntax-highlight"
 
 [tasks.benchmark-http-routing]
-description = "Run http-routing benchmark (wado serve vs Hono)"
+description = "Run http-routing benchmark (wado serve vs Hono on Node.js and Bun vs Axum)"
 run = "mise run -C benchmark http-routing"
 
 [tasks.report-wasm-size]
